Move ProcessService et al. to Mono.XBuild.Utilities and make it public
... to enable unit-testing.
