+++ mcs/class/Microsoft.Build.Utilities/Microsoft.Build.Utilities.dll.sources
@@ -6,10 +6,6 @@ Microsoft.Build.Utilities/AppDomainIsolatedTask.cs
 Microsoft.Build.Utilities/CommandLineBuilder.cs
 Microsoft.Build.Utilities/HostObjectInitializationStatus.cs
 Microsoft.Build.Utilities/Logger.cs
-Microsoft.Build.Utilities/IAsyncOperation.cs
-Microsoft.Build.Utilities/IProcessAsyncOperation.cs
-Microsoft.Build.Utilities/ProcessWrapper.cs
-Microsoft.Build.Utilities/ProcessService.cs
 Microsoft.Build.Utilities/TargetDotNetFrameworkVersion.cs
 Microsoft.Build.Utilities/Task.cs
 Microsoft.Build.Utilities/TaskItem.cs
