+++ mcs/class/Mono.XBuild.Utilities/Mono.XBuild.Utilities-net_2_0.csproj
@@ -47,12 +47,21 @@
     <Compile Include="..\..\tools\xbuild\XBuildConsts.cs">
       <Link>XBuildConsts.cs</Link>
     </Compile>
+    <Compile Include="..\System\System.Collections.Specialized\ProcessStringDictionary.cs">
+      <Link>ProcessStringDictionary.cs</Link>
+    </Compile>
     <Compile Include="Assembly\AssemblyInfo.cs" />
     <Compile Include="Mono.XBuild.BuildEngine\DirectoryScannerBase.cs" />
     <Compile Include="Mono.XBuild.BuildEngine\EventSource.cs" />
     <Compile Include="Mono.XBuild.BuildEngine\UnknownToolsVersionException.cs" />
+    <Compile Include="Mono.XBuild.Utilities\IAsyncOperation.cs" />
+    <Compile Include="Mono.XBuild.Utilities\IProcessAsyncOperation.cs" />
     <Compile Include="Mono.XBuild.Utilities\MSBuildUtils.cs" />
     <Compile Include="Mono.XBuild.Utilities\MonoLocationHelper.cs" />
+    <Compile Include="Mono.XBuild.Utilities\OperationHandler.cs" />
+    <Compile Include="Mono.XBuild.Utilities\OutWriter.cs" />
+    <Compile Include="Mono.XBuild.Utilities\ProcessService.cs" />
+    <Compile Include="Mono.XBuild.Utilities\ProcessWrapper.cs" />
     <Compile Include="Mono.XBuild.Utilities\ReservedNameUtils.cs" />
   </ItemGroup>
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
